worker/requirements.txt
# Computer Vision
ultralytics
Pillow
opencv-python

# Supabase
supabase
python-dotenv 